Extract ApplicationStartTests case
